46 refactor common code for later re-use
